refactor env variable script
